llava has been renamed to llava-hf, and llava-lht is now renamed to llava.
